[#930] Moved some classes
